[wip] Documentation improvement
The best place to start improving the docs would be the Readme, I particularly found it a little bit confusing.

I put it in wip because I'm not really happy with the order of things, the overall structure... and I would like to discuss it some more. Some things are subjective to every individual tastes, but this is a proposal, let's discuss things.